+++ CommonsLib_APP/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.2] - 2020-04-26
+
+### Improved BootstrapInitializers extensions collection.
+
 ## [1.0.1] - 2020-04-11
 
 ### Initial 1.0.1 Release